requirements.txt
flask==2.0.1
flask-cors==3.0.10
boto3==1.26.137
requests==2.28.2
PyPDF2==3.0.1
spacy==3.5.2
scikit-learn==1.2.2
numpy==1.24.3
python-dotenv==1.0.0
